requirements.txt
numpy
pandas
hyperspy
pyxem
tqdm
dask
matplotlib
diffsims
jupyter-slack
tensorflow
neptune-client
neptune-tensorflow-keras
pymongo
dnspython